MODHAADM-105: Unhandled exception when reading PURGE_LOGS_AFTER
https://folio-org.atlassian.net/browse/MODHAADM-105

Parsing the response may fail when doing "new JsonObject(response)", "getJsonArray" or "getJsonObject".

The parse error is not handled because the code is in "onSuccess(", therefor the request hangs forever.

Fix:

Use map, not onSuccess.

Avoid Promise that almost always is wrong and can be replaced with Future returning calls.

In addition fix CQL and percent encoding.